+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#1024](https://github.com/ClementTsang/bottom/pull/1024): Support FreeBSD temperature sensors based on `hw.temperature`.
 - [#1063](https://github.com/ClementTsang/bottom/pull/1063): Add buffer and cache memory tracking.
 - [#1106](https://github.com/ClementTsang/bottom/pull/1106): Add current battery charging state.
+- [#1115](https://github.com/ClementTsang/bottom/pull/1115): Add customizable process columns to config file.
 
 ## Changes